+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming version
 
+- Player can gain extra friendship points every whole hour while adventuring with companions (#109)
 - Mod's translation was moved to the separated folder `/locale/<localeCode>` and not requires the suffix ~~`.<localeCode>.json`~~
 - Old format content packs trigs an "unsafe" warning when they are loaded. (yellow only for debug mode)
 - Content pack in the new format (1.3) are safer.